+++ src/App.tsx
@@ -1,82 +1,54 @@
-import React from "react";
-import {
-  CustomOverlayMap,
-  CustomOverlayRoadview,
-  Map,
-  Roadview,
-} from "./lib/Kakaomap";
-
-const Content = () => (
-  <div className="overlay_info">
-    <a href="https://place.map.kakao.com/17600274">
-      <strong>월정리 해수욕장</strong>
-    </a>
-    <div className="desc">
-      <img
-        src="https://t1.daumcdn.net/localimg/localimages/07/mapapidoc/place_thumb.png"
-        alt=""
-      />
-      <span className="address">제주특별자치도 제주시 구좌읍 월정리 33-3</span>
-    </div>
-  </div>
-);
+import React, { useState } from "react";
+import { Map, MapMarker } from "./lib/Kakaomap";
 
 const SampleCreateMap = () => {
+  const [isOpen, setIsOpen] = useState(false);
+
   return (
-    <>
-      <Map // 로드뷰를 표시할 Container
-        center={{
-          // 지도의 중심좌표
-          lat: 33.5563,
-          lng: 126.79581,
-        }}
-        style={{
-          // 지도의 크기
-          width: "100%",
-          height: "300px",
-        }}
-        level={3}
-      >
-        <CustomOverlayMap
-          position={{
-            lat: 33.55635,
-            lng: 126.795841,
-          }}
-          options={{
-            xAnchor: 0.5,
-            yAnchor: 1.1,
-          }}
-        >
-          <Content />
-        </CustomOverlayMap>
-      </Map>
-      <Roadview // 로드뷰를 표시할 Container
+    <Map // 지도를 표시할 Container
+      center={{
+        // 지도의 중심좌표
+        lat: 33.450701,
+        lng: 126.570667,
+      }}
+      style={{
+        // 지도의 크기
+        width: "100%",
+        height: "450px",
+      }}
+      level={3} // 지도의 확대 레벨
+    >
+      <MapMarker // 인포윈도우를 생성하고 지도에 표시합니다
         position={{
-          // 지도의 중심좌표
-          lat: 33.5563,
-          lng: 126.7958,
-          radius: 50,
+          // 인포윈도우가 표시될 위치입니다
+          lat: 33.450701,
+          lng: 126.570667,
         }}
-        style={{
-          // 지도의 크기
-          width: "100%",
-          height: "300px",
-        }}
+        clickable={true} // 마커를 클릭했을 때 지도의 클릭 이벤트가 발생하지 않도록 설정합니다
+        onClick={() => setIsOpen(true)}
       >
-        <CustomOverlayRoadview
-          position={{
-            lat: 33.55635,
-            lng: 126.795841,
-          }}
-          options={{
-            xAnchor: 0.5,
-            yAnchor: 0.5,
-          }}
-        >
-          <Content />
-        </CustomOverlayRoadview>
-      </Roadview>
-    </>
+        {/* MapMarker의 자식을 넣어줌으로 해당 자식이 InfoWindow로 만들어지게 합니다 */}
+        {/* 인포윈도우에 표출될 내용으로 HTML 문자열이나 React Component가 가능합니다 */}
+        {isOpen && (
+          <div style={{ minWidth: "150px" }}>
+            <img
+              alt="close"
+              width="14"
+              height="13"
+              src="https://t1.daumcdn.net/localimg/localimages/07/mapjsapi/2x/bt_close.gif"
+              style={{
+                position: "absolute",
+                right: "5px",
+                top: "5px",
+                cursor: "pointer",
+              }}
+              onClick={() => setIsOpen(false)}
+            />
+            <div style={{ padding: "5px" }}>Hello World!</div>
+          </div>
+        )}
+      </MapMarker>
+    </Map>
   );
 };