+++ src/lib/Kakaomap/InfoWindow.tsx
@@ -1,60 +1,69 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useMemo, useRef } from "react";
 import ReactDom from "react-dom";
 
 interface InfoWindowProps {
   map: kakao.maps.Map | kakao.maps.Roadview;
   position: kakao.maps.LatLng | kakao.maps.Viewpoint;
-  options?: {
-    /**
-     * 인포윈도우를 열 때 지도가 자동으로 패닝하지 않을지의 여부 (기본값: false)
-     */
-    disableAutoPan?: boolean;
+  marker?: kakao.maps.Marker;
+  /**
+   * 인포윈도우를 열 때 지도가 자동으로 패닝하지 않을지의 여부 (기본값: false)
+   */
+  disableAutoPan?: boolean;
 
-    /**
-     * 삭제 가능한 인포윈도우
-     */
-    removable?: boolean;
+  /**
+   * 삭제 가능한 인포윈도우
+   */
+  removable?: boolean;
 
-    /**
-     * 인포윈도우 엘리먼트의 z-index 속성 값
-     */
-    zIndex?: number;
+  /**
+   * 인포윈도우 엘리먼트의 z-index 속성 값
+   */
+  zIndex?: number;
 
-    /**
-     * 로드뷰에 올라있는 인포윈도우의 높이 값(m 단위)
-     */
-    altitude?: number;
+  /**
+   * 로드뷰에 올라있는 인포윈도우의 높이 값(m 단위)
+   */
+  altitude?: number;
 
-    /**
-     * 로드뷰 상에서 인포윈도우의 가시반경(m 단위), 두 지점 사이의 거리가 지정한 값보다 멀어지면 인포윈도우는 보이지 않게 된다
-     */
-    range?: number;
-  };
+  /**
+   * 로드뷰 상에서 인포윈도우의 가시반경(m 단위), 두 지점 사이의 거리가 지정한 값보다 멀어지면 인포윈도우는 보이지 않게 된다
+   */
+  range?: number;
 }
 
 const InfoWindow: React.FC<InfoWindowProps> = ({
   map,
   position,
+  marker,
   children,
-  options,
+  altitude,
+  disableAutoPan,
+  range,
+  removable,
+  zIndex,
 }) => {
   const container = useRef(document.createElement("div"));
-  const [infoWindow, setInfoWindow] = useState<kakao.maps.InfoWindow>();
 
-  useEffect(() => {
-    const kakaoInfoWindow = new kakao.maps.InfoWindow({
-      ...options,
+  const infoWindow = useMemo(() => {
+    return new kakao.maps.InfoWindow({
+      altitude: altitude,
+      disableAutoPan: disableAutoPan,
+      range: range,
+      removable: removable,
+      zIndex: zIndex,
       content: container.current,
       position: position,
-      map: map,
     });
-    setInfoWindow(kakaoInfoWindow);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
+  useEffect(() => {
+    infoWindow.open(map, marker);
     return () => {
-      kakaoInfoWindow.close();
+      infoWindow.close();
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [map, container]);
+  }, [map, marker]);
 
   // 현재 위치에 표시되는 인포윈도우를 생성한다
   useEffect(() => {
@@ -63,19 +72,19 @@ const InfoWindow: React.FC<InfoWindowProps> = ({
   }, [infoWindow, position]);
 
   useEffect(() => {
-    if (!infoWindow || !options?.altitude) return;
-    infoWindow.setAltitude(options.altitude);
-  }, [infoWindow, options?.altitude]);
+    if (!infoWindow || !altitude) return;
+    infoWindow.setAltitude(altitude);
+  }, [infoWindow, altitude]);
 
   useEffect(() => {
-    if (!infoWindow || !options?.range) return;
-    infoWindow.setRange(options.range);
-  }, [infoWindow, options?.range]);
+    if (!infoWindow || !range) return;
+    infoWindow.setRange(range);
+  }, [infoWindow, range]);
 
   useEffect(() => {
-    if (!infoWindow || !options?.zIndex) return;
-    infoWindow.setZIndex(options.zIndex);
-  }, [infoWindow, options?.zIndex]);
+    if (!infoWindow || !zIndex) return;
+    infoWindow.setZIndex(zIndex);
+  }, [infoWindow, zIndex]);
 
   return ReactDom.createPortal(children, container.current);
 };