+++ src/lib/Kakaomap/hooks/index.tsx
@@ -1,4 +1,6 @@
 import useKakaoMapsSDK from "./useKakaoMapsSDK";
+import useMapSetCenter from "./useMapSetCenter";
+import useMapPanTo from "./useMapPanTo";
 import useRelayout from "./useRelayout";
 
-export { useKakaoMapsSDK, useRelayout };
+export { useKakaoMapsSDK, useMapSetCenter, useMapPanTo, useRelayout };